id-1: The First Version of the Game
The goal of this game is to escape the cave.

If the game starts, the stopwatch starts and you can see the wind-up toy. You should escape the cave by controlling the toy.

There are five levels and you can switch them by pressing num keys. The default level is level 5, so if you found an exit (which is located in the most bottom right place) it directly leads you to the message "You're finally out!" (but it's Korean). So, if you want to enjoy the whole level, press '1' and play from level 1.

Also, there isn't collisions between the wall and the toy (except for the outer walls) so try not to get hit by the wall. I will soon update the version with collisions!

Controls
- Use arrow keys to move the wind-up toy.
- Press number keys (1, 2, 3, 4, 5) to play the number of the level you want.